requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.70.0
vertexai==1.70.0
pydantic==2.9.2